Move url generation methods to their own interface
